requirements.txt
fastapi
uvicorn
pydantic-settings
openai
langchain
python-dotenv
ruff
ultralytics
opencv-python
Pillow